+++ src/main/java/io/github/astrapi69/entity/nameable/UniqueNameEntity.java
@@ -26,9 +26,9 @@ package io.github.astrapi69.entity.nameable;
 
 import java.io.Serializable;
 
-import javax.persistence.AttributeOverride;
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
+import jakarta.persistence.AttributeOverride;
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
 
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;