+++ src/main/java/io/github/astrapi69/entity/nameable/versionable/VersionableNameUUIDEntity.java
@@ -24,8 +24,8 @@
  */
 package io.github.astrapi69.entity.nameable.versionable;
 
-import javax.persistence.MappedSuperclass;
-import javax.persistence.Version;
+import jakarta.persistence.MappedSuperclass;
+import jakarta.persistence.Version;
 
 import lombok.AccessLevel;
 import lombok.Getter;