+++ src/main/java/io/github/astrapi69/entity/verifiable/BasicVerifiableEntity.java
@@ -26,7 +26,7 @@ package io.github.astrapi69.entity.verifiable;
 
 import java.io.Serializable;
 
-import javax.persistence.MappedSuperclass;
+import jakarta.persistence.MappedSuperclass;
 
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;